feat: 네비게이션 바 페이지 이동 기능 구현 (#9)
* TTOL-35 feat: 네비게이션 바 페이지 이동 기능 구현

Link 태그로 변경, 로그인/로그아웃 hover 시 아이콘 색상 변경되도록 추가

* TTOL-35 style: 로그인/로그아웃 페이지 텍스트 색상 수정

* TTOL-35 refactor: 필요없는 주석 제거

* TTOL-35 refactor: 메인 페이지를 대시보드 페이지로 변경

page.tsx 파일을 src/app 경로로 이동
